Extend table endpoint with wildcards support
Summary:
Wildcards add ability of mass table registering,deregistering and listing.

For list and del command any fields can contain "*" wildcard, which matches any
value of the corresponding field. Just ommiting the fields has the same effect
in the "list" commands, in "del" command this is not allowed to protect from
operator and software errors.

In addition to "*" wildcard, db and table fields of "add" command can contain
"%" wildcard, which has the same semantics as in MySQL.